Claude git it working, lol

+++ guest/firmware/mem/stack.c
@@ -7,8 +7,8 @@ uint64_t get_sp() {
 }
 
 void print_stack_usage() {
-    uint32_t sp = get_sp();
-    uint32_t used = 0x400000 - sp;
+    uint64_t sp = get_sp();
+    uint64_t used = 0x400000 - sp;
     serial_puts("stack: ");
     serial_putx(used);
     serial_puts(" Bytes used\n");
